+++ tests/index.js
@@ -166,7 +166,7 @@ const oneWeekAfterSummer = new DateTime(2021, 5, 28)
 
 describe('DateTime class', () => {
   test('extends Date', () => expect(summer).toBeInstanceOf(Date))
-  test('.getWeekNumber', () => expect(summer.getWeekNumber()).toBe(25))
-  test('.setWeekNumber', () => expect(summer.setWeekNumber(26)).toBe(oneWeekAfterSummer.getTime()))
+  test('.getWeek', () => expect(summer.getWeek()).toBe(25))
+  test('.setWeek', () => expect(summer.setWeek(26)).toBe(oneWeekAfterSummer.getTime()))
   test(".to('day')", () => expect(summer.to('day')).toBe('2021-06-28'))
 })