+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Compare with [last published version](https://github.com/meduzen/datetime-attrib
 
 ### New
 
-- Add a [`DateTime()` class](https://github.com/meduzen/datetime-attribute#the-datetime-class) extending the native `Date` object. Available methods: `getWeekNumber()`, `setWeekNumber()` and `to()`.
+- Add a [`DateTime()` class](https://github.com/meduzen/datetime-attribute#the-datetime-class) extending the native `Date` object. Available methods: `getWeek()`, `setWeek()` and `to()`.
 
 ### Fixed